refactor: 파일 구조 변경
테스트 코드를 작성하기 위한 파일구조로 변경
