+++ src/member/components/SignUpAdditionalRadio/index.jsx
@@ -1,0 +1,3 @@
+import SignUpAdditionalRadio from './SignUpAdditionalRadio';
+
+export default SignUpAdditionalRadio;
