+++ src/member/components/SignUpAddress/index.jsx
@@ -1,0 +1,3 @@
+import SignUpAddress from './SignUpAddress';
+
+export default SignUpAddress;
