+++ src/member/components/SignUpBasicInput/index.jsx
@@ -1,0 +1,3 @@
+import SignUpBasicInput from './SignUpBasicInput';
+
+export default SignUpBasicInput;
