+++ src/member/components/SignUpBirthInput/index.jsx
@@ -1,0 +1,3 @@
+import SignUpBirthInput from './SignUpBirthInput';
+
+export default SignUpBirthInput;
