+++ src/member/components/SignUpForm/index.jsx
@@ -1,0 +1,3 @@
+import SignUpForm from './SignUpForm';
+
+export default SignUpForm;
