+++ src/member/components/SignUpGenderRadio/index.jsx
@@ -1,0 +1,3 @@
+import SignUpGenderRadio from './SignUpGenderRadio';
+
+export default SignUpGenderRadio;
