+++ src/member/components/SignUpTermsAndConditions/index.jsx
@@ -1,0 +1,3 @@
+import SignUpTermsAndConditions from './SignUpTermsAndConditions';
+
+export default SignUpTermsAndConditions;
