+++ src/member/containers/SignUpFormContainer/SignUpFormContainer.jsx
@@ -1,4 +1,4 @@
-import SignUpForm from '../components/SignUpForm';
+import SignUpForm from '../../components/SignUpForm';
 
 export default function SignUpFormContainer() {
   return <SignUpForm />;
