+++ src/member/containers/SignUpFormContainer/index.jsx
@@ -1,0 +1,3 @@
+import SignUpFormContainer from './SignUpFormContainer';
+
+export default SignUpFormContainer;
